refactor: edit modal components + data fetching
