+++ package.json
@@ -21,6 +21,7 @@
     "react-draggable": "^4.4.5",
     "react-firebase-hooks": "^5.1.1",
     "react-icons": "^4.9.0",
+    "swr": "^2.2.2",
     "tailwindcss": "3.3.2"
   },
   "devDependencies": {
