+++ src/app/classroom/page.jsx
@@ -12,7 +12,7 @@ import StudentGrid from "@/components/democlass/student/StudentGrid"
 import TableGrid from "@/components/democlass/table/TableGrid"
 import Toolbar from "@/components/democlass/Toolbar"
 import Greeting from "@/components/democlass/Greeting"
-import Weather from "@/components/democlass/Weather"
+import Weather from "@/components/WeatherData"
 import Birthday from "@/components/democlass/Birthday"
 import Scribble from "@/components/Scribble"
 import Preloader from "@/components/Preloader"
