+++ src/components/LandingPage.jsx
@@ -8,6 +8,9 @@ import Scribble from "./Scribble"
 
 const LandingPage = () => {
   const [user, loading] = useAuthState(auth)
+  const creationTime = user?.metadata.creationTime
+  const lastSignInTime = user?.metadata.lastSignInTime
+
 
   const scribblesSvgs = [
     { src: '/assets/Scribbles/24.svg', className: 'absolute top-10 left-0 w-[75px] md:w-[100px] 2xl:w-[150px]' },
@@ -28,7 +31,9 @@ const LandingPage = () => {
         <h1 className="text-4xl md:text-6xl lg:text-8xl text-center font-bold uppercase tracking-wider text-yellow-400">Student Space</h1>
         <h2 className="order-[-1] text-sm lg:text-xl text-center text-secondaryTextClr pb-1">Elevate Classroom Engagement.</h2>
         <h2 className="md:text-lg lg:text-2xl text-center text-secondaryTextClr capitalize">Crafted exclusively for primary teachers.</h2>
-        <h3 className="text-2xl md:text-xl lg:text-2xl text-center pt-10 pb-2 text-secondaryTextClr">Get Started!</h3>
+        <h3 className="text-2xl md:text-xl lg:text-2xl text-center pt-10 pb-2 text-secondaryTextClr">
+          {!user ? 'Get Started!' : creationTime !== lastSignInTime ? 'Welcome Back!' : 'Get Started!'}
+        </h3>
         {user ? 
         <Link href={"/dashboard"} className="py-3 px-16 text-lg lg:text-xl bg-buttonClr text-primaryTextClr rounded-full hover:scale-105 duration-300">Dashboard</Link>
         : <Link href={"/login"} className="py-2 px-10 md:py-3 md:px-16 text-lg md:text-xl bg-buttonClr text-primaryTextClr rounded-full hover:scale-105 duration-300">Sign in</Link>}
